# Add new `Performance/RedundantStringChars` cop