Closes #139 

No need to call `#chars` and allocate an extra array when method can be directly called on string.

```ruby
# bad
str.chars[0..2]
str.chars.slice(0..2)

# good
str[0..2].chars

# bad
str.chars.first
str.chars.first(2)
str.chars.last
str.chars.last(2)

# good
str[0]
str[0...2].chars
str[-1]
str[-2..-1].chars

# bad
str.chars.take(2)
str.chars.drop(2)
str.chars.length
str.chars.size
str.chars.empty?

# good
str[0...2].chars
str[2..-1].chars
str.length
str.size
str.empty?
```

### Benchmark
```ruby
# frozen_string_literal: true

require 'bundler/inline'

gemfile(true) do
  gem 'benchmark-ips'
  gem 'benchmark-memory'
end

STR = 'a' * 20
RANGE = 10..15


puts('********* IPS *********')

Benchmark.ips do |x|
  x.report("String#chars[...]")   { STR.chars[RANGE] }
  x.report("String#[...]")        { STR[RANGE] }
  x.compare!
end

puts "********* MEMORY *********"

Benchmark.memory do |x|
  x.report("String#chars[...]")   { STR.chars[RANGE] }
  x.report("String#[...]")        { STR[RANGE] }
  x.compare!
end
```

### Results
```
********* IPS *********
Warming up --------------------------------------
   String#chars[...]    18.063k i/100ms
        String#[...]   214.729k i/100ms
Calculating -------------------------------------
   String#chars[...]    181.304k (± 3.5%) i/s -    921.213k in   5.087585s
        String#[...]      2.141M (± 2.2%) i/s -     10.736M in   5.016910s

Comparison:
        String#[...]:  2141163.0 i/s
   String#chars[...]:   181304.1 i/s - 11.81x  (± 0.00) slower

********* MEMORY *********
Calculating -------------------------------------
   String#chars[...]   920.000  memsize (     0.000  retained)
                        23.000  objects (     0.000  retained)
                         2.000  strings (     0.000  retained)
        String#[...]    40.000  memsize (     0.000  retained)
                         1.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)

Comparison:
        String#[...]:         40 allocated
   String#chars[...]:        920 allocated - 23.00x more
```