@fatkodima 2020-06-23
I thought about it a little more. This cop was originally created to primarily catch cases like `str.chars[0..2]` (and equivalent `str.chars.slice(0..2)`). This was originally implemented (incorrectly) and then dropped. I have added support to replace them with faster alternative, `str[0..2].chars`.

And also updated implementation to catch `#take(n)`, `#drop(n)`; `#first` and `#last` with argument.